V0.1.0/requirements.txt
# Web Framework & Server
fastapi
uvicorn

# Database (ORM & Async Driver)
tortoise-orm
asyncpg

# Configuration & Security
pydantic-settings
python-jose[cryptography]
bcrypt

# Validation & Utilities
pydantic[email] # 包含 email-validator
python-multipart
aiohttp